+++ lib/common/extensions/context_extension.dart
@@ -8,7 +8,14 @@ extension BuildContextExtension on BuildContext {
     required Widget content,
   }) {
     ScaffoldMessenger.of(this).showSnackBar(
-      SnackBar(content: content),
+      SnackBar(
+        behavior: SnackBarBehavior.floating,
+        margin: const EdgeInsets.symmetric(
+          horizontal: 12,
+          vertical: 5,
+        ),
+        content: content,
+      ),
     );
   }
