+++ lib/screens/main/main_screen.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+
+class MainScreen extends StatelessWidget {
+  const MainScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold(
+      body: SafeArea(
+        child: Center(
+          child: Text(
+            '메인화면',
+            style: TextStyle(fontSize: 50),
+          ),
+        ),
+      ),
+    );
+  }
+}
