+++ pubspec.yaml
@@ -22,6 +22,10 @@ dependencies:
 
   flex_color_scheme: ^7.3.1
   http:
+  shared_preferences: ^2.3.2
+  get: ^4.6.6
+  go_router: ^14.2.8
+
   # NOTE: UI/UX
 dev_dependencies:
   flutter_test:
